Refactor file structure for ease of testing
