--- bot_1/bot_saves_princess_runner.rb
@@ -1,10 +1,0 @@
-require './lib/grid'
-require './lib/bot_saves_princess'
-require './lib/game_play'
-
-def start
-  new_game = GamePlay.new
-  new_game.play
-end
-
-start
