+++ hacker_rank_solution_1.rb
