+++ hacker_rank_solution_2.rb
