+++ lib/bot_1/bot_saves_princess.rb
