+++ lib/bot_1/grid.rb
