+++ lib/bot_2/bot_saves_princess_2.rb
@@ -1,6 +1,6 @@
 require_relative 'grid_2'
 
-class BotSavesPrincess
+class BotSavesPrincess2
   attr_reader :grid
 
   def initialize(grid)
