+++ lib/bot_2/grid_2.rb
@@ -1,4 +1,4 @@
-class Grid
+class Grid2
   attr_reader :size, 
               :bot_location, 
               :grid,
