+++ spec/bot_2_spec/game_play_2_spec.rb
@@ -1,8 +1,8 @@
-require './lib/game_play_2' 
+require './lib/bot_2/game_play_2'  
 
-RSpec.describe GamePlay do
+RSpec.describe GamePlay2 do
   context 'build_grid' do
-    let(:game) {GamePlay.new}
+    let(:game) {GamePlay2.new}
 
     it 'can build a 3x3 grid' do
       expect(game.build_grid(3)).to eq(["---", "---", "---"])
@@ -14,7 +14,7 @@ RSpec.describe GamePlay do
   end
 
   context 'place_bot_in_grid' do
-    let(:game) {GamePlay.new}
+    let(:game) {GamePlay2.new}
 
     it 'can place the bot in the 5x5 grid' do
       expect(game.place_bot_in_grid(["-----", "-----", "-----", "-----", "-----"], 1, 4)).to eq(["-----", "----m", "-----", "-----", "-----"])
@@ -22,7 +22,7 @@ RSpec.describe GamePlay do
   end
 
   context 'place_princess_in_grid' do
-    let(:game) {GamePlay.new}
+    let(:game) {GamePlay2.new}
 
     it 'can place the princess in the grid' do
       expect(game.place_princess_in_grid(["-----", "----m", "-----", "-----", "-----"], 2, 2)).to eq(["-----", "----m", "--p--", "-----", "-----"])
