+++ spec/bot_2_spec/grid_2_spec.rb
@@ -1,11 +1,11 @@
-require './lib/grid_2' 
+require './lib/bot_2/grid_2' 
 
 RSpec.describe do 
   context 'initialize' do
-    let(:grid) {Grid.new(5, 2, 3, ["-----", "-----","p--m-", "-----", "-----"])}
+    let(:grid) {Grid2.new(5, 2, 3, ["-----", "-----","p--m-", "-----", "-----"])}
     
     it 'exists' do 
-      expect(grid).to be_a(Grid)
+      expect(grid).to be_a(Grid2)
     end
 
     it 'has a size attribute' do 
@@ -26,7 +26,7 @@ RSpec.describe do
   end
 
   context 'locate_character_location' do
-    let(:grid) {Grid.new(5, 2, 3, ["-----", "-----","p--m-", "-----", "-----"])}
+    let(:grid) {Grid2.new(5, 2, 3, ["-----", "-----","p--m-", "-----", "-----"])}
 
     it 'can locate the princess 5x5 grid' do
       expect(grid.locate_character_location("p")).to eq([2, 0])
